Implement basic hardcoded login
Change sync entities authentication from hardcoded token to a token generated from a hardcoded login. Add generate access token code, built into the sync code, which will be refactored out in a later phase. A "log in" button is added to the sync view temporarily to facilitate an easy log in action. This should later be implemented in a more robust fashion in the app.