Load build info from properties file
Instead of embedding various build environment variables, load it either from the system environment or from a keys.properties file. This is particularly useful for somewhat sensitive information, so it is not in the source code. This is not completely secure, as when the app is built, the sensitive information would be present in the app, which could be extracted from the app's resources.